Chilean 2021 First Round Presidential Election

This dataset contains the results of the first round of the 2021 Chilean presidential elections. It provides 9 possible voting options (7 candidates, blank, and null). Each ballot-box is identified by its id (BALLOT.BOX) and an electoral circumscription (ELECTORAL.DISTRICT). Additionally, it provides demographic information on voters' age range for each ballot box.

Usage

data("chile_election_2021")

Format

A data frame with 46,639 rows and 14 variables:

REGION

The region of the ELECTORAL.DISTRICT

ELECTORAL.DISTRICT

The electoral circumscription of the ballot box.

BALLOT.BOX

An identifier for the ballot box within the ELECTORAL.DISTRICT.

C1

The number of votes cast for the candidate Gabriel Boric.

C2

The number of votes cast for the candidate José Antonio Kast.

C3

The number of votes cast for the candidate Yasna Provoste.

C4

The number of votes cast for the candidate Sebastián Sichel.

C5

The number of votes cast for the candidate Eduardo Artés.

C6

The number of votes cast for the candidate Marco Enríquez-Ominami.

C7

The number of votes cast for the candidate Franco Parisi.

BLANK.VOTES

The number of blank votes.

NULL.VOTES

The number of null votes.

X18.19

Number of voters aged 18–19.

X20.29

Number of voters aged 20–29.

X30.39

Number of voters aged 30–39.

X40.49

Number of voters aged 40–49.

X50.59

Number of voters aged 50–59.

X60.69

Number of voters aged 60–69.

X70.79

Number of voters aged 70–79.

X80.

Number of voters aged 80 and older.

MISMATCH

Boolean that takes value TRUE if the ballot-box has a mismatch between the total number of votes and the total number of voters. If this is not the case, its value is FALSE.

Source

Chilean Electoral Service (SERVEL)

Examples

data("chile_election_2021")
head(chile_election_2021)
#>           REGION ELECTORAL.DISTRICT BALLOT.BOX C1 C2 C3 C4 C5 C6 C7 BLANK.VOTES
#> 1 DE ANTOFAGASTA  ANTOFAGASTA NORTE        126 12 22  3  4  1  5 71           0
#> 2 DE ANTOFAGASTA  ANTOFAGASTA NORTE        127 11 25  4  4  0  6 70           1
#> 3 DE ANTOFAGASTA  ANTOFAGASTA NORTE        128 16 23  3  3  1  5 70           1
#> 4 DE ANTOFAGASTA  ANTOFAGASTA NORTE        129 10 22  4  4  0  2 75           0
#> 5 DE ANTOFAGASTA  ANTOFAGASTA NORTE        130 16 23  4  3  1  3 76           0
#> 6 DE ANTOFAGASTA  ANTOFAGASTA NORTE        131 22 25  3  4  3  2 62           0
#>   NULL.VOTES X18.19 X20.29 X30.39 X40.49 X50.59 X60.69 X70.79 X80. MISMATCH
#> 1          0      1      7    103      3      0      3      1    0    FALSE
#> 2          0      2     10     98      6      4      1      0    0    FALSE
#> 3          2      3      5    111      2      2      0      1    0    FALSE
#> 4          2      2      2    103      7      3      2      0    0    FALSE
#> 5          1      4      9    112      1      2      0      0    0     TRUE
#> 6          3      3      8    103      4      3      1      0    2    FALSE